requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
langchain
langchain-community
PyGithub
faiss-cpu
sentence-transformers
loguru
pytest
pytest-asyncio
google-genai